집안일 카드 수정 완료
