+++ app/src/main/res/layout/fragment_main.xml
@@ -236,6 +236,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="3dp"
+                    android:layout_marginHorizontal="24dp"
                     android:orientation="vertical"
                     android:overScrollMode="never"
                     app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
@@ -245,7 +246,7 @@
                     app:layout_constraintTop_toBottomOf="@id/rv_week"
                     android:paddingBottom="76dp"
                     tools:itemCount="3"
-                    tools:listitem="@layout/item_section_recyclerview" />
+                    tools:listitem="@layout/item_housework" />
             </androidx.coordinatorlayout.widget.CoordinatorLayout>
         </androidx.constraintlayout.widget.ConstraintLayout>
         <include
