+++ app/src/main/res/values/strings.xml
@@ -79,6 +79,7 @@
     <string name="rule_place_holder">여기를 눌러 하우스의 규칙을 입력해주세요</string>
     <string name="main_today">오늘</string>
     <string name="main_complete">완료</string>
+    <string name="main_back_chore">되돌리기</string>
 
     <!-- dummy chores -->
     <string-array name="chore_array">
